pyproject.toml
[project]
name = "checkride-copilot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "mcp",
    "fastapi",
    "uvicorn",
    "httpx",
    "faster-whisper",
    "kokoro",
    "piper-tts",
    "nvidia-cublas-cu12; platform_system == 'Windows'",
    "nvidia-cudnn-cu12; platform_system == 'Windows'",
    "sounddevice",
    "pyaudio",
    "pynput",
    "chromadb",
    "sentence-transformers",
    "rank-bm25",
    "pypdf",
    "pydantic",
    "python-dotenv",
    "loguru",
    "pytest",
    "torch>=2.6.0",
    "torchvision",
    "torchaudio",
]

[project.optional-dependencies]
parakeet = [
    "nemo_toolkit[asr]>=2.7.0",
]

[project.scripts]
start = "orchestrator.orchestrator:main"
start-demo = "scripts.voice_demo:main"
start-demo-llm = "scripts.voice_spike_llm_first:main"
test = "pytest:main"

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }
torchaudio = { index = "pytorch-cu128" }

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = [
    "collector*",
    "diagnostic*",
    "mcp_server*",
    "mission*",
    "orchestrator*",
    "retrieval*",
    "voice*",
]



